Former-commit-id: 42adf625d9266d4dfff74b74d78c1a560057d0d3

+++ ifs-data-service/src/main/java/org/innovateuk/ifs/competition/domain/CompetitionFunder.java
@@ -2,7 +2,7 @@ package org.innovateuk.ifs.competition.domain;
 
 
 import javax.persistence.*;
-import java.math.BigDecimal;
+import java.math.BigInteger;
 
 /**
  * Entity model to store the Competition Co-Funders.
@@ -19,7 +19,7 @@ public class CompetitionFunder {
     private Competition competition;
 
     private String funder;
-    private BigDecimal funderBudget;
+    private BigInteger funderBudget;
     private Boolean coFunder;
 
     public Long getId() {
@@ -46,11 +46,11 @@ public class CompetitionFunder {
         this.funder = funder;
     }
 
-    public BigDecimal getFunderBudget() {
+    public BigInteger getFunderBudget() {
         return funderBudget;
     }
 
-    public void setFunderBudget(BigDecimal funderBudget) {
+    public void setFunderBudget(BigInteger funderBudget) {
         this.funderBudget = funderBudget;
     }
